# feat(strings): faster string length and repeat