--- a/src/internals/strings/impl/length.ts
+++ b/src/internals/strings/impl/length.ts
@@ -20,11 +20,11 @@ type LengthUp<
           : never
       ) extends infer $DoubleC extends string
       ? `$${T}` extends `${$DoubleC}${infer $Rest}`
-        ? $Cache["length"] extends 12 // 2^12 is the last block size within the complexity limit
+        ? $Cache["length"] extends 12 // 2^13 is the last block size within the complexity limit
           ? LengthDownSlow<
               $Rest,
               Add<$Acc["length"], Mul<C[1]["length"], 2>>,
-              $Cache
+              [[$DoubleC, [...C[1], ...C[1]]], ...$Cache]
             >
           : LengthUp<
               $Rest,