--- a/src/internals/strings/impl/strings.ts
+++ b/src/internals/strings/impl/strings.ts
@@ -3,3 +3,4 @@ export * from "./trim";
 export * from "./replace";
 export * from "./repeat";
 export * from "./compare";
+export * from "./length";