This optimization makes use of a self-memoizing cache in the string length algorithm which allows it to run on a logarithmic time scale. This allows us to get the length of strings that are over 800,000 characters long performantly without freezing the language server.