@ecyrbe 2023-02-15
Not sure of the improvement here, current implementation handles 999 characters size just well :
```ts
    type chars999 = Call<Strings.Repeat<999>, 'a'>;
    type length999 = Call<Strings.Length, res100>;
```
Can you measure timings  for strings of these sizes :  
- 5
- 10
- 100
- 250
- 999

to be sure this is not everkill