@sno2 2023-02-15
> Not sure of the improvement here, current implementation handles 999 characters size just well and maybe handling 9999 is too much for practical use cases :
> 
> ```ts
>     type chars999 = Call<Strings.Repeat<999>, 'a'>;
>     type length999 = Call<Strings.Length, res100>;
> ```
> 
> Can you measure timings for strings of these sizes :
> 
> * 5
> * 10
> * 100
> * 250
> * 999
> * 9999 (will only work with you impl, not current one)
> 
> to be sure this is not overkill for lower sized strings

When considering the type system, the way that I see the order that we should optimize should be that runtime complexity trumps all other aspects. We only have a few thousand iterations before the type system fails us. Therefore, using a logarithmic implementation will make it possible to do much more trickery with string types.

I did try benchmarking this with `Hello, world!`. However, it seems to be negligible performance differences because hyperfine found that the after implementation was faster after running for 1.5 minutes on the same programs.

(Before)
  Range (min … max):    9.647 s … 15.403 s    10 runs

(After)
  Range (min … max):    9.200 s … 11.496 s    10 runs

The performance is definitely better for strings larger than 16 because it takes log2(x) plus a few more for all iterations. Here are the benchmarks for 1,000 characters.

(Before)
  Time (mean ± σ):     11.915 s ±  2.519 s    [User: 22.778 s, System: 2.058 s]
  Range (min … max):    9.053 s … 15.845 s    10 runs
  
(After)
  Time (mean ± σ):      8.297 s ±  0.774 s    [User: 16.844 s, System: 1.418 s]
  Range (min … max):    6.764 s …  9.350 s    10 runs
  
I also improved the algorithm even more such that it can take above 800,000 characters by transitioning into a linear algorithm on the 2^13 pattern after it gets that large.