@sno2 2023-02-15
@gvergnaud Would it be alright if I apply this optimization to other string methods in this PR as well? If I applied the logarithmic cache idea to `Repeat`, then we could repeat ~400,000 characters and then get the length in one go.