@sno2 2023-02-15
Actually, `Repeat` would probably be the only other method I would implement it for. We could do it for string trimming, but not really sure if it's worth it.