@ecyrbe 2023-02-15
LGTM, thank you for the tests.
  
Yes, go ahead and you can do Repeat also

If you have other perfs improvements in other areas, they are welcomed.

Numbers are already pretty optimised since they are using a lot of lookup tables.

If you think you can optimise string comparison, go ahead also