@sno2 2023-02-16
Alright, this is ready for review. I am very pleased with the result of the `Repeat` methods because there is not a cap on the size of a string literal so we can scale the patterns to 2^n indefinitely and create strings millions of characters long with ease :^)

Here is a little round-trip example of using the two features together:

```ts
import { Strings, Pipe } from ".";

type A = Pipe<"a", [Strings.Repeat<840_000>, Strings.Length]>;
//   ^? 840000
```