@ecyrbe 2023-02-16
@sno2 
here is an improved, more readable version i implemented that does the same (supports more than 1_000_00 repeats like yours)
```ts
export type Repeat<
  T extends string,
  N extends number,
  Acc extends string = "",
  Calc extends {
    Quotient: number;
    Remainder: number;
  } = DivMod<N, 2>
> = N extends 0
  ? Acc
  : N extends 1
  ? `${Acc}${T}`
  : Calc["Remainder"] extends 0
  ? Repeat<RepeatX2<T>, Calc["Quotient"], Acc>
  : Repeat<T, Sub<N, 1>, `${Acc}${T}`>;
```
it reads almost the same as the js version:
```ts

const repeat = (s: string, count: number, acc: string = ''): string => {
  if (count === 0) {
    return acc;
  } else if (count % 2 === 0) {
    return repeat(s + s, count / 2, acc);
  } else {
    return repeat(s, count - 1, acc + s);
  }
}
```

Do you want to also simplify the Length one ? else i'll do it, because it's really hard to read and maintain